aws-sagemaker/local_train/requirements.txt
sagemaker
sagemaker[local]
tensorflow==2.7.0
Pillow